requirements.txt
python-telegram-bot==21.5
sqlalchemy==2.0.31
psycopg2-binary==2.9.9
alembic==1.13.2
cryptography==42.0.8
solders==0.21.0
solana==0.34.3
python-dotenv==1.0.1
httpx==0.27.0
